FLINK-2435: Refactoring of CsvReader tests
