+++ flink-tests/src/test/java/org/apache/flink/test/io/csv/custom/type/NestedCustomJsonTypeStringParser.java
@@ -18,9 +18,10 @@
 
 package org.apache.flink.test.io.csv.custom.type;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.types.parser.FieldParser;
 import org.apache.flink.types.parser.StringParser;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
