+++ flink-tests/src/test/java/org/apache/flink/test/io/csv/custom/type/complex/GenericsAwareCustomJsonTypeStringParser.java
@@ -18,10 +18,11 @@
 
 package org.apache.flink.test.io.csv.custom.type.complex;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.types.parser.FieldParser;
 import org.apache.flink.types.parser.StringParser;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
