+++ flink-tests/src/test/java/org/apache/flink/test/io/csv/custom/type/simple/SimpleCustomJsonTypeStringParser.java
@@ -18,9 +18,10 @@
 
 package org.apache.flink.test.io.csv.custom.type.simple;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.types.parser.FieldParser;
 import org.apache.flink.types.parser.StringParser;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
